all: use `int()` for string conversion
In Py2 `int(string)` automatically returns `long` if the `string` is too long.
There is only `int` type in Py3. Hence, using `int()` for string conversion is
both correct and version independent.

+++ widgets/device_description_settings.py
@@ -130,7 +130,7 @@ class DeviceDescriptionSettingsWidget(QOMDescriptionSettingsWidget):
 
             if val_type is int:
                 try:
-                    new_val = long(new_val, 0)
+                    new_val = int(new_val, 0)
                 except:
                     
                     w.config(bg = "red")
